:truck: Rename FacebookMessenger resources
